bug: adding save data for Building manager

+++ Assets/Scripts/Models/BuildingEntry.cs
@@ -33,7 +33,7 @@ public class BuildingEntry
 
     public long MaxCardNum;
 
-    public int ProbabilityOfBeingAttacked;
+    public double ProbabilityOfBeingAttacked;
 
     public int HeatDissipationLevel;
